+++ StabilityMatrix.Avalonia/ViewModels/CheckpointBrowser/CivitAiBrowserViewModel.cs
@@ -153,8 +153,6 @@ public sealed partial class CivitAiBrowserViewModel : TabViewModelBase, IInfinit
 
         EventManager.Instance.NavigateAndFindCivitModelRequested += OnNavigateAndFindCivitModelRequested;
 
-        var settingsSelectedBaseModels = settingsManager.Settings.SelectedCivitBaseModels;
-
         var filterPredicate = Observable
             .FromEventPattern<PropertyChangedEventArgs>(this, nameof(PropertyChanged))
             .Where(x =>
@@ -199,7 +197,7 @@ public sealed partial class CivitAiBrowserViewModel : TabViewModelBase, IInfinit
                 .Transform(baseModel => new BaseModelOptionViewModel
                 {
                     ModelType = baseModel,
-                    IsSelected = settingsSelectedBaseModels.Contains(baseModel),
+                    IsSelected = settingsManager.Settings.SelectedCivitBaseModels.Contains(baseModel),
                 })
                 .SortAndBind(
                     AllBaseModels,
@@ -225,6 +223,7 @@ public sealed partial class CivitAiBrowserViewModel : TabViewModelBase, IInfinit
 
         var settingsTransactionObservable = this.WhenPropertyChanged(x => x.SelectedBaseModels)
             .Throttle(TimeSpan.FromMilliseconds(50))
+            .Skip(1)
             .ObserveOn(SynchronizationContext.Current)
             .Subscribe(_ =>
             {
@@ -360,7 +359,7 @@ public sealed partial class CivitAiBrowserViewModel : TabViewModelBase, IInfinit
         }
 
         dontSearch = true;
-        baseModelCache.Edit(updater => updater.Load(baseModels));
+        baseModelCache.EditDiff(baseModels, static (a, b) => a.Equals(b, StringComparison.OrdinalIgnoreCase));
         dontSearch = false;
     }
 
@@ -415,6 +414,9 @@ public sealed partial class CivitAiBrowserViewModel : TabViewModelBase, IInfinit
         var timer = Stopwatch.StartNew();
         var queryText = request.Query;
         var models = new List<CivitModel>();
+        // Store original request for caching
+        var originalRequestStr = JsonSerializer.Serialize(request);
+
         CivitModelsResponse? modelsResponse = null;
         try
         {
@@ -428,6 +430,7 @@ public sealed partial class CivitAiBrowserViewModel : TabViewModelBase, IInfinit
                     foreach (var chunk in idChunks)
                     {
                         request.CommaSeparatedModelIds = string.Join(",", chunk);
+                        request.Limit = 100;
                         var chunkModelsResponse = await civitApi.GetModels(request);
 
                         if (chunkModelsResponse.Items != null)
@@ -485,10 +488,12 @@ public sealed partial class CivitAiBrowserViewModel : TabViewModelBase, IInfinit
                 // Add to database
                 await liteDbContext.UpsertCivitModelAsync(models);
                 // Add as cache entry
+
+                var originalRequest = JsonSerializer.Deserialize<CivitModelsRequest>(originalRequestStr);
                 cacheNew = await liteDbContext.UpsertCivitModelQueryCacheEntryAsync(
                     new CivitModelQueryCacheEntry
                     {
-                        Id = ObjectHash.GetMd5Guid(request),
+                        Id = ObjectHash.GetMd5Guid(originalRequest),
                         InsertedAt = DateTimeOffset.UtcNow,
                         Request = request,
                         Items = models,
@@ -626,7 +631,6 @@ public sealed partial class CivitAiBrowserViewModel : TabViewModelBase, IInfinit
             Nsfw = "true", // Handled by local view filter
             Sort = SortMode,
             Period = SelectedPeriod,
-            Limit = 30,
         };
 
         if (NextPageCursor != null)