+++ StabilityMatrix.Avalonia/ViewModels/Dialogs/ModelVersionViewModel.cs
@@ -1,24 +1,20 @@
-﻿using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Linq;
-using CommunityToolkit.Mvvm.ComponentModel;
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+using StabilityMatrix.Avalonia.ViewModels.Base;
+using StabilityMatrix.Core.Helper;
 using StabilityMatrix.Core.Models.Api;
 using StabilityMatrix.Core.Services;
 
 namespace StabilityMatrix.Avalonia.ViewModels.Dialogs;
 
-public partial class ModelVersionViewModel : ObservableObject
+public partial class ModelVersionViewModel : DisposableViewModelBase
 {
     private readonly IModelIndexService modelIndexService;
 
     [ObservableProperty]
-    private CivitModelVersion modelVersion;
+    public partial CivitModelVersion ModelVersion { get; set; }
 
     [ObservableProperty]
-    private ObservableCollection<CivitFileViewModel> civitFileViewModels;
-
-    [ObservableProperty]
-    private bool isInstalled;
+    public partial bool IsInstalled { get; set; }
 
     public ModelVersionViewModel(IModelIndexService modelIndexService, CivitModelVersion modelVersion)
     {
@@ -32,11 +28,7 @@ public partial class ModelVersionViewModel : ObservableObject
                 && modelIndexService.ModelIndexBlake3Hashes.Contains(file.Hashes.BLAKE3)
             ) ?? false;
 
-        CivitFileViewModels = new ObservableCollection<CivitFileViewModel>(
-            (
-                ModelVersion.Files?.Select(file => new CivitFileViewModel(modelIndexService, file)) ?? []
-            ).OrderBy(a => a.CivitFile.Type == CivitFileType.TrainingData ? 1 : 0)
-        );
+        EventManager.Instance.ModelIndexChanged += ModelIndexChanged;
     }
 
     public void RefreshInstallStatus()
@@ -46,5 +38,19 @@ public partial class ModelVersionViewModel : ObservableObject
                 file is { Type: CivitFileType.Model, Hashes.BLAKE3: not null }
                 && modelIndexService.ModelIndexBlake3Hashes.Contains(file.Hashes.BLAKE3)
             ) ?? false;
+    }
+
+    private void ModelIndexChanged(object? sender, EventArgs e)
+    {
+        RefreshInstallStatus();
+    }
+
+    protected override void Dispose(bool disposing)
+    {
+        if (disposing)
+        {
+            EventManager.Instance.ModelIndexChanged -= ModelIndexChanged;
+        }
+        base.Dispose(disposing);
     }
 }